Access control

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,16 @@ use App\Models\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -69,6 +79,9 @@ class PostsController extends Controller
      */
     public function edit(Post $post)
     {
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorised Access');
+        }
         return view('posts.edit', compact('post'));
     }
 
@@ -101,6 +114,9 @@ class PostsController extends Controller
      */
     public function destroy(Post $post)
     {
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorised Access');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post deleted successfully');
     }
